Fix Java only builds
Model was getting applied in all cases, which is incorrect
