+++ src/main/groovy/edu/wpi/first/gradlerio/wpi/WPIPlugin.groovy
@@ -3,6 +3,7 @@ package edu.wpi.first.gradlerio.wpi
 import edu.wpi.first.gradlerio.wpi.dependencies.WPIJavaDeps
 import edu.wpi.first.gradlerio.wpi.dependencies.WPIJsonDepsPlugin
 import edu.wpi.first.gradlerio.wpi.dependencies.WPINativeDeps
+import edu.wpi.first.gradlerio.wpi.dependencies.WPINativeJsonDepRules
 import edu.wpi.first.gradlerio.wpi.dependencies.tools.WPIToolsPlugin
 import edu.wpi.first.gradlerio.wpi.toolchain.WPIToolchainPlugin
 import edu.wpi.first.vscode.GradleVsCode
@@ -32,6 +33,7 @@ class WPIPlugin implements Plugin<Project> {
             project.pluginManager.apply(WPINativeDeps)
             project.pluginManager.apply(WPIToolchainPlugin)
             project.pluginManager.apply(GradleVsCode)
+            project.pluginManager.apply(WPINativeJsonDepRules)
         }
 
         project.tasks.register("wpi") { Task task ->
