+++ src/main/groovy/edu/wpi/first/gradlerio/wpi/dependencies/WPIJsonDepsPlugin.groovy
@@ -106,7 +106,6 @@ class WPIJsonDepsPlugin implements Plugin<Project> {
 
     @Override
     void apply(Project project) {
-        project.pluginManager.apply(NativeDepsPlugin)
         def wpi = project.extensions.getByType(WPIExtension)
 
         def jsonExtension = project.extensions.create('wpiJsonDeps', WPIJsonDepsExtension, project)
@@ -137,8 +136,6 @@ class WPIJsonDepsPlugin implements Plugin<Project> {
                 }
             }
         }
-
-        DependencySpecExtension dse = project.extensions.getByType(DependencySpecExtension)
 
         // Add all URLs from dependencies
         jsonExtension.dependencies.each { JsonDependency dep ->
@@ -215,25 +212,15 @@ class WPIJsonDepsPlugin implements Plugin<Project> {
         })
 
         project.extensions.add('useCppVendorLibraries', { Object closureArg, String... ignoreLibraries ->
-            if (closureArg in TargetedNativeComponent) {
-                TargetedNativeComponent component = (TargetedNativeComponent)closureArg
-                component.binaries.withType(NativeBinarySpec).all { NativeBinarySpec bin ->
-                    Set<DelegatedDependencySet> dds = []
-                    jsonExtension.dependencies.find { (!ignoreLibraries.contains(it.name) && !ignoreLibraries.contains(it.uuid)) }.each { JsonDependency dep ->
-                        dep.cppDependencies.collect { CppArtifact art ->
-                            dds << new DelegatedDependencySet(dep.uuid + art.libName, bin, dse, art.skipOnUnknownClassifier)
-                        }
-                    }
+            useCppVendorLibraries(project, jsonExtension, closureArg, ignoreLibraries)
+        })
+    }
 
-                    bin.inputs.withType(DependentSourceSet) { DependentSourceSet dss ->
-                        dds.each { DelegatedDependencySet set ->
-                            dss.lib(set)
-                        }
-                    }
-
-                }
-            } else if (closureArg in NativeBinarySpec) {
-                NativeBinarySpec bin = (NativeBinarySpec) closureArg
+    private void useCppVendorLibraries(Project project, WPIJsonDepsExtension jsonExtension, Object closureArg, String... ignoreLibraries) {
+        def dse = project.extensions.getByType(DependencySpecExtension)
+        if (closureArg in TargetedNativeComponent) {
+            TargetedNativeComponent component = (TargetedNativeComponent)closureArg
+            component.binaries.withType(NativeBinarySpec).all { NativeBinarySpec bin ->
                 Set<DelegatedDependencySet> dds = []
                 jsonExtension.dependencies.find { (!ignoreLibraries.contains(it.name) && !ignoreLibraries.contains(it.uuid)) }.each { JsonDependency dep ->
                     dep.cppDependencies.collect { CppArtifact art ->
@@ -246,122 +233,24 @@ class WPIJsonDepsPlugin implements Plugin<Project> {
                         dss.lib(set)
                     }
                 }
-            } else {
-                throw new GradleException('Unknown type for useVendorLibraries target. You put this declaration in a weird place.')
+
             }
-        })
-    }
-
-    static class WPIJsonDepRules extends RuleSource {
-        @Mutate
-        void addJsonLibraries(NativeDepsSpec libs, final ExtensionContainer extensionContainer) {
-            def wpi = extensionContainer.getByType(WPIExtension)
-            def jsonExtension = extensionContainer.getByType(WPIJsonDepsExtension)
-            def common = { NativeLib lib ->
-                lib.targetPlatforms = ['roborio']
-                lib.headerDirs = []
-                lib.sourceDirs = []
-                lib.staticMatchers = []
-                lib.sharedMatchers = []
-                lib.dynamicMatchers = []
-            }
-
-            def nativeclassifier = (
-                    OperatingSystem.current().isWindows() ?
-                            System.getProperty("os.arch") == 'amd64' ? 'windowsx86-64' : 'windowsx86' :
-                            OperatingSystem.current().isMacOsX() ? "osxx86-64" :
-                                    OperatingSystem.current().isLinux() ? "linuxx86-64" :
-                                            null
-            )
-
-            jsonExtension.dependencies.each { JsonDependency dep ->
-                dep.cppDependencies.each { CppArtifact art ->
-                    def name = dep.uuid + art.libName
-                    def supportNative = art.validClassifiers.contains(nativeclassifier)
-                    def supportAthena = art.validClassifiers.contains('linuxathena')
-                    def mavenBase = "${art.groupId}:${art.artifactId}:${art.version}"
-                    def cfgName = "native_${name}"
-                    libs.create("${name}_headers", NativeLib) { NativeLib lib ->
-                        common(lib)
-                        if (supportNative)
-                            lib.targetPlatforms << 'desktop'
-                        lib.headerDirs << ''
-                        lib.maven = "${mavenBase}:${art.headerClassifier}@zip"
-                        lib.configuration = cfgName
-                        null
-                    }
-
-                    if (art.isHeaderOnly) {
-                        // Create header only lib
-                        libs.create(name, CombinedNativeLib) { CombinedNativeLib lib ->
-                            lib.libs << "${name}_headers".toString()
-                            lib.targetPlatforms = ['roborio']
-                            if (supportNative)
-                                lib.targetPlatforms << 'desktop'
-                            null
-                        }
-                        return
-                    }
-
-                    if (supportAthena) {
-                        libs.create("${name}_athena", NativeLib) { NativeLib lib ->
-                            common(lib)
-                            lib.libraryName = "${name}_binaries"
-                            if (art.sharedLibrary) {
-                                lib.sharedMatchers = ["**/lib${art.libName}.so".toString()]
-                                lib.dynamicMatchers = lib.sharedMatchers
-                            } else {
-                                lib.staticMatchers = ["**/static/*.a".toString()]
-                            }
-                            lib.maven = "${mavenBase}:linuxathena@zip"
-                            lib.configuration = cfgName
-                            null
-                        }
-                    }
-
-                    if (art.hasSources) {
-                        libs.create("${name}_sources", NativeLib) { NativeLib lib ->
-                            common(lib)
-                            if (supportNative)
-                                lib.targetPlatforms << 'desktop'
-                            lib.sourceDirs << ''
-                            lib.maven = "${mavenBase}:${art.sourcesClassifier}@zip"
-                            lib.configuration = cfgName
-                            null
-                        }
-                    }
-
-                    if (supportNative && nativeclassifier != null) {
-                        libs.create("${name}_native", NativeLib) { NativeLib lib ->
-                            common(lib)
-                            lib.libraryName = "${name}_binaries"
-                            lib.targetPlatforms = ['desktop']
-                            if (art.sharedLibrary) {
-                                lib.staticMatchers = ["**/*${art.libName}.lib".toString()]
-                                lib.sharedMatchers = ["**/*${art.libName}.so".toString(), "**/*${art.libName}.dylib".toString()]
-
-                                lib.dynamicMatchers = lib.sharedMatchers + "**/${art.libName}.dll".toString()
-                            } else {
-                                lib.staticMatchers = ["**/*${art.libName}.lib".toString(), "**/*${art.libName}.so".toString()]
-                            }
-                            lib.maven = "${mavenBase}:${nativeclassifier}@zip"
-                            lib.configuration = "${cfgName}_desktop"
-                            null
-                        }
-                    }
-
-                    libs.create(name, CombinedNativeLib) { CombinedNativeLib lib ->
-                        lib.libs << "${name}_binaries".toString() << "${name}_headers".toString()
-                        if (art.hasSources) {
-                            lib.libs << "${name}_sources".toString()
-                        }
-                        lib.targetPlatforms = ['roborio']
-                        if (supportNative)
-                            lib.targetPlatforms << 'desktop'
-                        null
-                    }
+        } else if (closureArg in NativeBinarySpec) {
+            NativeBinarySpec bin = (NativeBinarySpec) closureArg
+            Set<DelegatedDependencySet> dds = []
+            jsonExtension.dependencies.find { (!ignoreLibraries.contains(it.name) && !ignoreLibraries.contains(it.uuid)) }.each { JsonDependency dep ->
+                dep.cppDependencies.collect { CppArtifact art ->
+                    dds << new DelegatedDependencySet(dep.uuid + art.libName, bin, dse, art.skipOnUnknownClassifier)
                 }
             }
+
+            bin.inputs.withType(DependentSourceSet) { DependentSourceSet dss ->
+                dds.each { DelegatedDependencySet set ->
+                    dss.lib(set)
+                }
+            }
+        } else {
+            throw new GradleException('Unknown type for useVendorLibraries target. You put this declaration in a weird place.')
         }
     }
 }
